+++ packages/cli/src/workspace-packages.d.ts
@@ -18,6 +18,8 @@ declare module "@zachariahredfield/playbook-node" {
 declare module "@zachariahredfield/playbook-engine" {
   export const loadConfig: (...args: any[]) => Promise<any>;
   export const generateRepositoryHealth: (...args: any[]) => any;
+  export const buildRepoAdoptionReadiness: (...args: any[]) => any;
+  export type RepoAdoptionReadiness = any;
   export type ArtifactHygieneReport = any;
   export const analyzePullRequest: (...args: any[]) => any;
   export const formatAnalyzePrGithubComment: (...args: any[]) => string;